разбил файл main.js на модули
